Special:ApprovedRevs now works as well as original, but...
The three functions of Special:ApprovedRevs are now working:
1) Show all pages with an approved revision
2) Show pages whose approved revision is not their latest
3) Show unapproved pages

(1) and (2) can either be setup to show all pages within the
approved_revs table, or only those pages that are approvable
per MediaWiki:Approvedrevs-permissions. I'm not sure which
direction is best, yet.

A fourth function is yet to be added:
4) Show all pages with approvals that are not considered approvable

(4) would show pages that are within the approved_revs table,
but are not considered approvable per MediaWiki:Approvedrevs-permissions.

ALSO, all four of these functions need to be duplicated for
the approved_revs_files table.