Add execute() method to query page classes to modify title
SpecialApprovedRevsQueryPage and SpecialApprovedFilesQueryPage required an execute() method to be able to modify the special page titles using the OutputPage setPageTitle() method. Good i18n messages were added to en.json.

Removed code using QueryPage::doQuery, since that is deprecated as of MW 1.18 (ApprovedRevs supports MW 1.19+) and the setPageTitle() method is within the execute() method, not the doQuery() method.