Always purge cache after approving rev (breaking change for SMW 1.9.0)
This change makes Approved Revs incompatible with SMW 1.9.0, but no
other SMW versions.

Removed restriction to only run $article->doPurge() in SMW < 1.9,
which was causing cache issues with history pages. When revisions
were marked approved, navigating to the history page did not show
the just-approved revision as approved. The previously approved
revision still showed approved (or no approval, if that was the
case).

The reason for not performing doPurge in SMW > 1.9 was because of an
SMW bug in v1.9.0, which was fixed in v1.9.1. Approved Revs should
not maintain workarounds for SMW bugs that:
 * Have been fixed
 * Only affected one minor release
 * Negatively impact Approved Revs

See commits e80ac09f and c5370dd4 for more info.